Update build script

+++ avm/client/build_wasm.sh
@@ -1,9 +1,13 @@
-#!/bin/sh
-(
-  cd ../../air-interpreter;
+#! /usr/bin/env bash
+
+if [[ -f dist/avm.wasm ]]; then
+  echo "air-interpreter wasm binary already present"
+  exit 0
+else
+  cd ../../air-interpreter
   marine build  --release --features marine
-)
+  cd -
 
-mkdir -p ./dist/
-cp ../../target/wasm32-wasi/release/air_interpreter_server.wasm ./dist/avm.wasm
-
+  mkdir -p ./dist/
+  cp ../../target/wasm32-wasi/release/air_interpreter_server.wasm ./dist/avm.wasm
+fi
